requirements.txt
streamlit>=1.2.0
numpy>=1.21.2
tensorflow-cpu>=2.6.0
matplotlib>=3.4.3
Pillow>=8.4.0
pandas>=1.3.4
setuptools>=58.0.4
tensorflow-hub>=0.12.0
opencv-python-headless